+++ src/DependencyInjection/Compiler/InitializerPass.php
@@ -11,9 +11,9 @@ use Symfony\Component\DependencyInjection\Reference;
  *
  * @author David Buchmann <mail@davidbu.ch>
  */
-class InitializerPass implements CompilerPassInterface
+final class InitializerPass implements CompilerPassInterface
 {
-    public function process(ContainerBuilder $container)
+    public function process(ContainerBuilder $container): void
     {
         if (!$container->hasDefinition('doctrine_phpcr.initializer_manager')) {
             return;