+++ src/ManagerRegistry.php
@@ -2,16 +2,16 @@
 
 namespace Doctrine\Bundle\PHPCRBundle;
 
+use Doctrine\ODM\PHPCR\DocumentManagerInterface;
 use Doctrine\ODM\PHPCR\PHPCRException;
+use PHPCR\SessionInterface;
 use Psr\Container\ContainerInterface;
 use Symfony\Bridge\Doctrine\ManagerRegistry as BaseManagerRegistry;
 
 /**
  * Symfony aware manager registry.
- *
- * @internal this class is intended to be used as service, but not for code reuse
  */
-class ManagerRegistry extends BaseManagerRegistry
+final class ManagerRegistry extends BaseManagerRegistry implements ManagerRegistryInterface
 {
     public function __construct(
         ContainerInterface $container,
@@ -38,11 +38,9 @@ class ManagerRegistry extends BaseManagerRegistry
      *
      * @param string $alias
      *
-     * @return string
-     *
      * @throws PHPCRException
      */
-    public function getAliasNamespace($alias)
+    public function getAliasNamespace($alias): string
     {
         foreach (array_keys($this->getManagers()) as $name) {
             try {
@@ -54,14 +52,30 @@ class ManagerRegistry extends BaseManagerRegistry
         throw PHPCRException::unknownDocumentNamespace($alias);
     }
 
+    public function getManager($name = null): DocumentManagerInterface
+    {
+        return parent::getManager($name);
+    }
+
+    public function resetManager($name = null): DocumentManagerInterface
+    {
+        return parent::resetManager($name);
+    }
+
+    public function getManagerForClass($class = null): ?DocumentManagerInterface
+    {
+        return parent::getManagerForClass($class);
+    }
+
+    public function getConnection($name = null): SessionInterface
+    {
+        return parent::getConnection($name);
+    }
+
     /**
      * Get the admin connection associated to the connection.
-     *
-     * @param string $name
-     *
-     * @return object
      */
-    public function getAdminConnection($name = null)
+    public function getAdminConnection(?string $name = null): SessionInterface
     {
         if (null === $name) {
             $name = $this->getDefaultConnectionName();