+++ tests/Fixtures/App/Controller/TestController.php
@@ -2,15 +2,16 @@
 
 namespace Doctrine\Bundle\PHPCRBundle\Tests\Fixtures\App\Controller;
 
-use Doctrine\Bundle\PHPCRBundle\ManagerRegistry;
+use Doctrine\Bundle\PHPCRBundle\ManagerRegistryInterface;
 use Doctrine\Bundle\PHPCRBundle\Tests\Fixtures\App\Document\ReferrerDocument;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Response;
 
 class TestController extends AbstractController
 {
-    private $registry;
+    private ManagerRegistryInterface $registry;
 
-    public function __construct(ManagerRegistry $registry)
+    public function __construct(ManagerRegistryInterface $registry)
     {
         $this->registry = $registry;
     }
@@ -18,7 +19,7 @@ class TestController extends AbstractController
     /**
      * Do a simple PHPCR request.
      */
-    public function phpcrRequest()
+    public function phpcrRequest(): Response
     {
         $dm = $this->registry->getManager();